refactor: Match databases dir to namespace.
