+++ test/orbitdb-custom-database-types.test.js
@@ -1,7 +1,7 @@
 import { strictEqual, deepStrictEqual, notStrictEqual } from 'assert'
 import rmrf from 'rimraf'
 import * as IPFS from 'ipfs-core'
-import { getDatabaseType } from '../src/db/index.js'
+import { getDatabaseType } from '../src/databases/index.js'
 import { OrbitDB, addDatabaseType, Database } from '../src/index.js'
 import config from './config.js'
